refactor: Reorganize the core module
- Migrated all utility functions to new `utils` module
- Moved all constants from core module to new `constants` module
